refactor: move model related file to model_executor dir
