+++ pipefuser/model_executor/models/transformers/pixart_transformer_2d.py
@@ -8,8 +8,8 @@ from diffusers.models.embeddings import PatchEmbed
 from diffusers.models.transformers.transformer_2d import Transformer2DModelOutput
 from diffusers.utils import is_torch_version
 
-from pipefuser.base_wrapper import PipeFuserBaseWrapper
-from pipefuser.models.transformers import PipeFuserTransformerBaseWrapper
+from pipefuser.model_executor.base_wrapper import PipeFuserBaseWrapper
+from pipefuser.model_executor.models.transformers import PipeFuserTransformerBaseWrapper
 from .register import PipeFuserTransformerWrappersRegister
 from pipefuser.config import ParallelConfig, RuntimeConfig
 from pipefuser.distributed import (
